Move scoped CSS handler registration
Is this the correct place?
