+++ packages/runtime-common/realm.ts
@@ -69,6 +69,7 @@ import type * as CardAPI from 'https://cardstack.com/base/card-api';
 import { createResponse } from './create-response';
 import { mergeRelationships } from './merge-relationships';
 import type { LoaderType } from 'https://cardstack.com/base/card-api';
+import { decodeScopedCSSRequest, isScopedCSSRequest } from 'glimmer-scoped-css';
 
 export type RealmInfo = {
   name: string;
@@ -156,6 +157,7 @@ export class Realm {
     this.#useTestingDomain = Boolean(opts?.useTestingDomain);
     this.#loader = loader;
     this.#loader.registerURLHandler(this.maybeHandle.bind(this));
+    this.#loader.registerURLHandler(maybeHandleScopedCSSRequest);
     this.#adapter = adapter;
     this.#searchIndex = new SearchIndex(
       this,
@@ -983,4 +985,26 @@ function sseToChunkData(type: string, data: string, id?: string): string {
     info.push(`id: ${id}`);
   }
   return info.join('\n') + '\n\n';
+}
+
+const isFastBoot = typeof (globalThis as any).FastBoot !== 'undefined';
+
+async function maybeHandleScopedCSSRequest(req: Request) {
+  if (isScopedCSSRequest(req.url)) {
+    if (isFastBoot) {
+      return Promise.resolve(new Response('// skipped scoped CSS'));
+    } else {
+      let decodedCSS = decodeScopedCSSRequest(req.url);
+      return Promise.resolve(
+        new Response(`
+          let styleNode = document.createElement('style');
+          let styleText = document.createTextNode(\`${decodedCSS}\`);
+          styleNode.appendChild(styleText);
+          document.body.appendChild(styleNode);
+        `)
+      );
+    }
+  } else {
+    return Promise.resolve(null);
+  }
 }
